+++ aepp/flowservice.py
@@ -1660,12 +1660,12 @@ class FlowManager:
             patchOperation = [{'op': 'replace',
             'path': f'/transformations/{myIndex}',
             'value': {'name': 'Mapping',
-            'params': {'mappingId': operation['params']['mappingId'],
-            'mappingVersion': operation['params']['mappingVersion']}}}
+            'params': {'mappingId': operation['mappingId'],
+            'mappingVersion': operation['mappingVersion']}}}
             ]
         else:
             raise Exception('Could not find a mapping transformation in the flow')
-        res = self.updateFlow(self.id,self.etag,patchOperation)
+        res = self.updateFlow(patchOperation)
         self.flowData = res
         self.__setAttributes__(res)
         return res